Rebased Serve/Core benchmarks to use common harness
