+++ python/ray/serve/_private/benchmarks/streaming/streaming_handle_throughput.py
@@ -6,59 +6,22 @@ import click
 
 from ray import serve
 from ray.serve._private.benchmarks.common import run_throughput_benchmark
+from ray.serve._private.benchmarks.streaming.common import Endpoint, Caller, IOMode
 from ray.serve.handle import DeploymentHandle, RayServeHandle
 
 
 @serve.deployment(ray_actor_options={"num_cpus": 0})
-class Downstream:
-    def __init__(self, tokens_per_request: int):
-        logging.getLogger("ray.serve").setLevel(logging.WARNING)
-
-        self._tokens_per_request = tokens_per_request
-
-    def stream(self):
-        for i in range(self._tokens_per_request):
-            yield "hi"
+class EndpointDeployment(Endpoint):
+    pass
 
 
 @serve.deployment
-class Caller:
-    def __init__(
-        self,
-        downstream: RayServeHandle,
-        *,
-        tokens_per_request: int,
-        batch_size: int,
-        num_trials: int,
-        trial_runtime: float,
-    ):
-        logging.getLogger("ray.serve").setLevel(logging.WARNING)
-
-        self._h: DeploymentHandle = downstream.options(
-            use_new_handle_api=True,
-            stream=True,
-        )
-        self._tokens_per_request = tokens_per_request
-        self._batch_size = batch_size
-        self._num_trials = num_trials
-        self._trial_runtime = trial_runtime
+class CallerDeployment(Caller):
 
     async def _consume_single_stream(self):
-        async for _ in self._h.stream.remote():
-            pass
-
-    async def _do_single_batch(self):
-        await asyncio.gather(
-            *[self._consume_single_stream() for _ in range(self._batch_size)]
-        )
-
-    async def run_benchmark(self) -> Tuple[float, float]:
-        return await run_throughput_benchmark(
-            fn=self._do_single_batch,
-            multiplier=self._batch_size * self._tokens_per_request,
-            num_trials=self._num_trials,
-            trial_runtime=self._trial_runtime,
-        )
+        async for r in self._h.stream.remote():
+            # Blackhole the response
+            self.sink(r)
 
 
 @click.command(help="Benchmark streaming deployment handle throughput.")
@@ -92,15 +55,23 @@ class Caller:
     default=5,
     help="Duration to run each trial of the benchmark for (seconds).",
 )
+@click.option(
+    "--io-mode",
+    type=str,
+    default="async",
+    help="Controls mode of the streaming generation (either 'sync' or 'async')",
+)
 def main(
     tokens_per_request: int,
     batch_size: int,
     num_replicas: int,
     num_trials: int,
     trial_runtime: float,
+    io_mode: str,
 ):
-    app = Caller.bind(
-        Downstream.options(num_replicas=num_replicas).bind(tokens_per_request),
+    app = CallerDeployment.bind(
+        EndpointDeployment.options(num_replicas=num_replicas).bind(tokens_per_request),
+        mode=IOMode(io_mode.upper()),
         tokens_per_request=tokens_per_request,
         batch_size=batch_size,
         num_trials=num_trials,
@@ -112,7 +83,8 @@ def main(
 
     mean, stddev = h.run_benchmark.remote().result()
     print(
-        "DeploymentHandle streaming throughput {}: {} +- {} tokens/s".format(
+        "DeploymentHandle streaming throughput ({}) {}: {} +- {} tokens/s".format(
+            io_mode.upper(),
             f"(num_replicas={num_replicas}, "
             f"tokens_per_request={tokens_per_request}, "
             f"batch_size={batch_size})",
